requirements.txt
scikit-learn
fastapi
pydantic
uvicorn
gunicorn
odoorpc
jupyterlab
dspy-ai
groq
python-dotenv
chromadb
transitions
python-jose
passlib
requests
celery
sqlalchemy
redis